win32 빌드 컴파일러 경고 제거

+++ Plugins/Funapi/Source/Funapi/Private/google/protobuf/descriptor.cc
@@ -149,7 +149,7 @@ string ToCamelCase(const string& input) {
   string result;
   result.reserve(input.size());
 
-  for (int i = 0; i < input.size(); i++) {
+  for (int i = 0; i < (int)input.size(); i++) {
     if (input[i] == '_') {
       capitalize_next = true;
     } else if (capitalize_next) {
@@ -587,7 +587,7 @@ DescriptorPool::Tables::~Tables() {
   // Note that the deletion order is important, since the destructors of some
   // messages may refer to objects in allocations_.
   STLDeleteElements(&messages_);
-  for (int i = 0; i < allocations_.size(); i++) {
+  for (int i = 0; i < (int)allocations_.size(); i++) {
     operator delete(allocations_[i]);
   }
   STLDeleteElements(&strings_);
@@ -628,17 +628,17 @@ void DescriptorPool::Tables::RollbackToLastCheckpoint() {
   const CheckPoint& checkpoint = checkpoints_.back();
 
   for (int i = checkpoint.pending_symbols_before_checkpoint;
-       i < symbols_after_checkpoint_.size();
+       i < (int)symbols_after_checkpoint_.size();
        i++) {
     symbols_by_name_.erase(symbols_after_checkpoint_[i]);
   }
   for (int i = checkpoint.pending_files_before_checkpoint;
-       i < files_after_checkpoint_.size();
+       i < (int)files_after_checkpoint_.size();
        i++) {
     files_by_name_.erase(files_after_checkpoint_[i]);
   }
   for (int i = checkpoint.pending_extensions_before_checkpoint;
-       i < extensions_after_checkpoint_.size();
+       i < (int)extensions_after_checkpoint_.size();
        i++) {
     extensions_.erase(extensions_after_checkpoint_[i]);
   }
@@ -658,7 +658,7 @@ void DescriptorPool::Tables::RollbackToLastCheckpoint() {
       file_tables_.begin() + checkpoint.file_tables_before_checkpoint,
       file_tables_.end());
   for (int i = checkpoint.allocations_before_checkpoint;
-       i < allocations_.size();
+       i < (int)allocations_.size();
        i++) {
     operator delete(allocations_[i]);
   }
@@ -1155,7 +1155,7 @@ void DescriptorPool::FindAllExtensions(
     vector<int> numbers;
     if (fallback_database_->FindAllExtensionNumbers(extendee->full_name(),
                                                     &numbers)) {
-      for (int i = 0; i < numbers.size(); ++i) {
+      for (int i = 0; i < (int)numbers.size(); ++i) {
         int number = numbers[i];
         if (tables_->FindExtension(extendee, number) == NULL) {
           TryFindExtensionInFallbackDatabase(extendee, number);
@@ -1744,7 +1744,7 @@ bool RetrieveOptions(int depth,
   const Reflection* reflection = options.GetReflection();
   vector<const FieldDescriptor*> fields;
   reflection->ListFields(options, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (int i = 0; i < (int)fields.size(); i++) {
     int count = 1;
     bool repeated = false;
     if (fields[i]->is_repeated()) {
@@ -1794,7 +1794,7 @@ bool FormatLineOptions(int depth, const Message &options, string *output) {
   string prefix(depth * 2, ' ');
   vector<string> all_options;
   if (RetrieveOptions(depth, options, &all_options)) {
-    for (int i = 0; i < all_options.size(); i++) {
+    for (int i = 0; i < (int)all_options.size(); i++) {
       strings::SubstituteAndAppend(output, "$0option $1;\n",
                                    prefix, all_options[i]);
     }
@@ -3155,7 +3155,7 @@ void DescriptorBuilder::ValidateSymbolName(
     AddError(full_name, proto, DescriptorPool::ErrorCollector::NAME,
              "Missing name.");
   } else {
-    for (int i = 0; i < name.size(); i++) {
+    for (int i = 0; i < (int)name.size(); i++) {
       // I don't trust isalnum() due to locales.  :(
       if ((name[i] < 'a' || 'z' < name[i]) &&
           (name[i] < 'A' || 'Z' < name[i]) &&
@@ -3171,7 +3171,7 @@ void DescriptorBuilder::ValidateSymbolName(
 bool DescriptorBuilder::ValidateQualifiedName(const string& name) {
   bool last_was_period = false;
 
-  for (int i = 0; i < name.size(); i++) {
+  for (int i = 0; i < (int)name.size(); i++) {
     // I don't trust isalnum() due to locales.  :(
     if (('a' <= name[i] && name[i] <= 'z') ||
         ('A' <= name[i] && name[i] <= 'Z') ||
@@ -3252,7 +3252,7 @@ template<class DescriptorT> void DescriptorBuilder::AllocateOptionsImpl(
 void DescriptorBuilder::AddRecursiveImportError(
     const FileDescriptorProto& proto, int from_here) {
   string error_message("File recursively imports itself: ");
-  for (int i = from_here; i < tables_->pending_files_.size(); i++) {
+  for (int i = from_here; i < (int)tables_->pending_files_.size(); i++) {
     error_message.append(tables_->pending_files_[i]);
     error_message.append(" -> ");
   }
@@ -3317,7 +3317,7 @@ const FileDescriptor* DescriptorBuilder::BuildFile(
   //   mid-file, but that's pretty ugly, and I'm pretty sure there are
   //   some languages out there that do not allow recursive dependencies
   //   at all.
-  for (int i = 0; i < tables_->pending_files_.size(); i++) {
+  for (int i = 0; i < (int)tables_->pending_files_.size(); i++) {
     if (tables_->pending_files_[i] == proto.name()) {
       AddRecursiveImportError(proto, i);
       return NULL;
