+++ Source/funapi_plugin_ue4/Tests/libprotobuf/LibProtobufTest.cpp
@@ -1665,12 +1665,12 @@ bool FFunapiLibProtobufUnknownFieldSetTest::RunTest(const FString& Parameters)
   auto CheckDeleteByNumber = [](const std::vector<int>& field_numbers, int deleted_number,
     const std::vector<int>& expected_field_nubmers) {
     google::protobuf::UnknownFieldSet unknown_fields;
-    for (int i = 0; i < field_numbers.size(); ++i) {
+    for (int i = 0; i < (int)field_numbers.size(); ++i) {
       unknown_fields.AddFixed32(field_numbers[i], i);
     }
     unknown_fields.DeleteByNumber(deleted_number);
     verify(expected_field_nubmers.size() == unknown_fields.field_count());
-    for (int i = 0; i < expected_field_nubmers.size(); ++i) {
+    for (int i = 0; i < (int)expected_field_nubmers.size(); ++i) {
       verify(expected_field_nubmers[i] ==
         unknown_fields.field(i).number());
     }
